feat: move next path utils out of the core lib
