+++ .changeset/clean-phones-push.md
@@ -1,0 +1,6 @@
+---
+'@ts-rest/core': patch
+'@ts-rest/next': patch
+---
+
+Move next lib utils out of @ts-rest/core to reduce bundle size
