+++ libs/ts-rest/core/src/index.ts
@@ -1,9 +1,7 @@
-export * from './lib/dsl';
 export * from './lib/client';
-export * from './lib/server-utils';
-export * from './lib/type-utils';
-export * from './lib/server-utils';
-export * from './lib/zod-utils';
-export * from './lib/status-codes';
+export * from './lib/dsl';
 export * from './lib/paths';
 export * from './lib/query';
+export * from './lib/status-codes';
+export * from './lib/type-utils';
+export * from './lib/zod-utils';
