+++ libs/ts-rest/next/src/lib/ts-rest-next.spec.ts
@@ -1,7 +1,6 @@
 import { initContract } from '@ts-rest/core';
-import { NextApiResponse } from 'next';
+import { NextApiRequest, NextApiResponse } from 'next';
 import { createNextRoute, createNextRouter } from './ts-rest-next';
-import { mockReq } from './test-helpers';
 
 const c = initContract();
 
@@ -150,3 +149,25 @@ describe('createNextRouter', () => {
     });
   });
 });
+
+export const mockReq = (
+  url: string,
+  args: {
+    query?: Record<string, string>;
+    body?: unknown;
+    method: string;
+  }
+): NextApiRequest => {
+  const paramArray = url.split('/').splice(1);
+
+  const req = {
+    query: {
+      ...args.query,
+      ['ts-rest']: paramArray,
+    },
+    body: args.body,
+    method: args.method,
+  } as unknown as NextApiRequest;
+
+  return req;
+};
